move fixed journal's init_sync to sync module
